Use brokered service for IManagedHotReloadService in VS
