+++ src/EditorFeatures/Core/EditAndContinue/ManagedHotReloadServiceProxy.cs
@@ -7,6 +7,7 @@ using System.Collections.Immutable;
 using System.Composition;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.CodeAnalysis.BrokeredServices;
 using Microsoft.CodeAnalysis.Host.Mef;
 using Microsoft.VisualStudio.Debugger.Contracts.HotReload;
 using InternalContracts = Microsoft.CodeAnalysis.Contracts.EditAndContinue;
@@ -17,18 +18,25 @@ namespace Microsoft.CodeAnalysis.EditAndContinue;
 [Export(typeof(InternalContracts.IManagedHotReloadService))]
 [method: ImportingConstructor]
 [method: Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
-internal sealed class ManagedHotReloadServiceProxy(Lazy<IManagedHotReloadService> service) :
+internal sealed class ManagedHotReloadServiceProxy(IServiceBrokerProvider serviceBrokerProvider) :
+    BrokeredServiceProxy<IManagedHotReloadService>(serviceBrokerProvider.ServiceBroker, BrokeredServiceDescriptors.DebuggerManagedHotReloadService),
     InternalContracts.IManagedHotReloadService
 {
-    public async ValueTask<ImmutableArray<InternalContracts.ManagedActiveStatementDebugInfo>> GetActiveStatementsAsync(CancellationToken cancellation)
-        => (await service.Value.GetActiveStatementsAsync(cancellation).ConfigureAwait(false)).SelectAsArray(a => a.ToContract());
+    public async ValueTask<ImmutableArray<InternalContracts.ManagedActiveStatementDebugInfo>> GetActiveStatementsAsync(CancellationToken cancellationToken)
+    {
+        var result = await InvokeAsync((service, cancellationToken) => service.GetActiveStatementsAsync(cancellationToken), cancellationToken).ConfigureAwait(false);
+        return result.SelectAsArray(info => info.ToContract());
+    }
 
-    public async ValueTask<InternalContracts.ManagedHotReloadAvailability> GetAvailabilityAsync(Guid module, CancellationToken cancellation)
-        => (await service.Value.GetAvailabilityAsync(module, cancellation).ConfigureAwait(false)).ToContract();
+    public async ValueTask<InternalContracts.ManagedHotReloadAvailability> GetAvailabilityAsync(Guid module, CancellationToken cancellationToken)
+    {
+        var result = await InvokeAsync((service, module, cancellationToken) => service.GetAvailabilityAsync(module, cancellationToken), module, cancellationToken).ConfigureAwait(false);
+        return result.ToContract();
+    }
 
-    public ValueTask<ImmutableArray<string>> GetCapabilitiesAsync(CancellationToken cancellation)
-        => service.Value.GetCapabilitiesAsync(cancellation);
+    public ValueTask<ImmutableArray<string>> GetCapabilitiesAsync(CancellationToken cancellationToken)
+        => InvokeAsync((service, cancellationToken) => service.GetCapabilitiesAsync(cancellationToken), cancellationToken);
 
-    public ValueTask PrepareModuleForUpdateAsync(Guid module, CancellationToken cancellation)
-        => service.Value.PrepareModuleForUpdateAsync(module, cancellation);
+    public ValueTask PrepareModuleForUpdateAsync(Guid module, CancellationToken cancellationToken)
+        => InvokeAsync((service, module, cancellationToken) => service.PrepareModuleForUpdateAsync(module, cancellationToken), module, cancellationToken);
 }
