+++ src/VisualStudio/DevKit/Impl/Microsoft.VisualStudio.LanguageServices.DevKit.csproj
@@ -17,6 +17,7 @@
     <Compile Include="..\..\..\EditorFeatures\Core\EditAndContinue\Contracts\ContractWrappers.cs" Link="EditAndContinue\ContractWrappers.cs" />
     <Compile Include="..\..\..\EditorFeatures\Core\EditAndContinue\ManagedHotReloadLanguageService.cs" Link="EditAndContinue\ManagedHotReloadLanguageService.cs" />
     <Compile Include="..\..\..\EditorFeatures\Core\EditAndContinue\ManagedHotReloadLanguageServiceImpl.cs" Link="EditAndContinue\ManagedHotReloadLanguageServiceImpl.cs" />
+    <Compile Include="..\..\..\EditorFeatures\Core\EditAndContinue\ManagedHotReloadServiceProxy.cs" Link="EditAndContinue\ManagedHotReloadServiceProxy.cs" />
     <Compile Include="..\..\..\Workspaces\Remote\ServiceHub\Services\EditAndContinue\EditAndContinueLogReporter.cs" Link="EditAndContinue\EditAndContinueLogReporter.cs" />
     <Compile Include="..\..\..\Workspaces\Remote\ServiceHub\Services\EditAndContinue\HotReloadLoggerProxy.cs" Link="EditAndContinue\HotReloadLoggerProxy.cs" />
     <Compile Include="..\..\..\VisualStudio\Core\Def\Telemetry\Shared\*.cs" LinkBase="Logging" />
